Fix : 기본 프사 삭제 현상 수정

+++ orury-domain/src/main/java/org/orury/domain/post/infrastructure/ImageStoreImpl.java
@@ -60,7 +60,7 @@ public class ImageStoreImpl implements ImageStore {
 
     @Override
     public void delete(String profile) {
-        if (StringUtils.isBlank(profile)) return;
+        if (StringUtils.isBlank(profile) || defaultImage.equals(profile)) return;
         var link = ImageUtil.splitUrlToImage(profile);
         amazonS3.deleteObject(bucket + USER.getName(), link);
     }
